Refactor chapter.js file to chapterManager for modularization
